rename mapData to map_data
