+++ kcauto.sikuli/combat.sikuli/map_data.py
